Worked on the version system

+++ src/main/java/net/rafael/plugins/creeper/recover/update/PluginVersion.java
@@ -33,7 +33,7 @@ public class PluginVersion {
     }
 
     public int compare(PluginVersion pluginVersion) {
-        return Double.compare(pluginVersion.asInt(), this.asInt());
+        return Integer.compare(pluginVersion.asInt(), this.asInt());
     }
 
     public int asInt() {
